# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: S2S 本地供需撮合平台 API
  version: 1.0.0-batch1
  description: |
    ## 1. 基础约定

    | 项 | 口径 |
    | --- | --- |
    | 基址 | `https://<自有域名>/api/v1` |
    | 时间格式 | RFC3339 UTC，例 `2026-09-30T12:00:00Z` |
    | 坐标系 | GCJ-02（火星坐标），经纬度均保留 5 位小数 |
    | 分页 | `page` 从 1 起，`page_size` 默认 20、上限 50 |
    | 压缩 | 响应强制 gzip |
    | 文件上传 | 无 multipart，统一 OSS 两步式直传（ticket + commit） |

    ## 2. 统一响应包

    ```json
    { "code": 0, "message": "ok", "data": { }, "request_id": "req_7f3a..." }
    ```

    - `code`：业务码，`0` 为成功，非 0 见错误码表；**前 3 位与 HTTP 状态码对齐**
    - `message`：中文文案，非 0 时可直接呈现给用户
    - `data`：成功时为业务负载，失败时恒为 `null`
    - `request_id`：服务端生成的请求标识，是 UI 报错时唯一回显值

    ## 3. 错误码表（25 个 code：成功码 1 个 + 错误码 24 个）

    | code | HTTP | 含义 | 用户可见 | Retry-After |
    | --- | --- | --- | --- | --- |
    | 0 | 200 | 成功 | - | - |
    | 40001 | 400 | 参数缺失或非法 | 是 | - |
    | 40002 | 400 | 未勾选用户协议与隐私政策 | 是 | - |
    | 40101 | 401 | 未登录或登录态失效 | 是（静默续期） | - |
    | 40105 | 401 | 验证码连续错误，登录锁定 15 分钟 | 是 | 是 |
    | 40301 | 403 | 未实名，今日详情浏览额度已用尽 | 是 | - |
    | 40302 | 403 | 高敏类目需先完成资质认证 | 是 | - |
    | 40303 | 403 | 该类目当前禁止发布 | 是 | - |
    | 40304 | 403 | 未实名用户发布数量已达上限 | 是 | - |
    | 40305 | 403 | 无权限访问该资源（后台 RBAC 越权，含读与写） | 是（**后台专用**，App 端不返回） | - |
    | 40901 | 409 | 内容命中敏感词 | 是 | - |
    | 40902 | 409 | 图片未通过审核 | 是 | - |
    | 40903 | 409 | 数据已被其他操作修改（乐观锁冲突） | 是 | - |
    | 41001 | 410 | 帖子已下架或已过期 | 是 | - |
    | 42901 | 429 | AI 生成配额已用尽 | 是 | 是 |
...

...

    ## 4. 客户端错误码行为四分类

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /posts/precheck:
    post:
      tags: [post]
      summary: 发布前置校验（不落库）
      operationId: precheckPost
      x-batch: Batch1
      x-domain: post
      description: |
        在用户点击「发布」之前提前暴露阻断项，避免填完长表单才报错。
        校验项与 `POST /posts` 完全一致但**不写库**：
        敏感词（`40901`）、图片过审（`40902`）、类目禁发（`40303`）、
        高敏类目资质（`40302`）、未实名发布上限（`40304`）。

        校验通过时 `data.passed=true`；发现问题时**仍返回 200 + code=0**，
        由 `data.blocks[]` 列出全部阻断项（一次性给全，不逐条打断用户）。
      parameters:
        - $ref: '#/components/parameters/XInteractionId'
        - $ref: '#/components/parameters/XDeviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostDraft'
      responses:
        '200':
          description: 校验完成（通过或存在阻断项均为 200/code=0）
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PrecheckResult'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /posts:
    post:
      tags: [post]
      summary: 发布帖子（资源或需求）
      operationId: createPost
      x-batch: Batch1
      x-domain: post
      description: |
        ### 落库派生字段（客户端不可传，传了也被忽略）

        | 字段 | 派生方式 |
        | --- | --- |
        | `l2_category_id` | STORED 生成列 `leaf_category_id DIV 100` |
        | `completeness_level` | STORED 生成列，由 `required_full` / `address_precise` / `leaf_matched` 三条件达成数映射：3→2、2→1、其他→0 |
        | `grid_id` | 服务端按 0.0045° 步长算法计算 |
        | `expire_at` | 发布时间 + 有效天数（见 lib/nfr_constants.dart NfrPostLifecycle.validDays） |
        | `version` | 初始 0 |

        ### 联系方式加密

        `contact_value` 落库为双列范式：
        `contact_hash`（HMAC-SHA256 + pepper，`BINARY(32)`，承载等值查询）+
        `contact_value_enc`（AES-GCM-256 + 随机 IV，`VARBINARY`，AAD 绑定 `post_id`）。
        **本接口响应中不回显 `contact_value`**；完整联系方式只由
        `GET /posts/{id}/contact` 返回。

        成功后异步产出埋点事件 `post_published`。
      parameters:
        - $ref: '#/components/parameters/XInteractionId'
        - $ref: '#/components/parameters/XDeviceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCreateRequest'
      responses:
        '200':
          description: 发布成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PostDetail'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /posts/{post_id}:
    get:
      tags: [post]
      summary: 获取帖子详情
      operationId: getPostDetail
      x-batch: Batch1
      x-domain: post
      security: []
      description: |
        ### 未登录可只读访问（口径以 PRD §12.3 为准）

        本接口 `security: []`，未登录可只读查看详情。**注意与
        `/posts/{id}/contact` 区分**：后者未登录一律 `40101`，本接口不是。

        ### 未登录详情浏览限频（`42907`）

        因 PRD §14.3 已裁定**不做坐标模糊化**，本接口返回的 `address`
        含门牌号，若未登录不限频则精确住址可零成本批量采集。故未登录访问
        按**设备维每日 ≤30 次、IP 维每日 ≤100 次**（沿用 §7.7 联系方式的
        设备/IP 维阈值，非新增数字），任一超限返 `42907` + `Retry-After`。
        已登录用户不受此限。

        ### 出参必带 `version`

        用于 `PATCH /posts/{id}/status` 的乐观锁入参。

        ### 未实名详情浏览额度（`40301`）

        PRD §12.3 规定已登录未实名用户每日放行 3 条、第 4 条回 `40301`。

        **Batch1 该额度关闭**：由 `system_config.detail_quota_enabled` 开关控制，
        Batch1 置 `off`。原因是 `cert` 域整体延后导致全员
        `realname_status='none'`，额度生效会把「地图→详情→联系」P0 闭环卡死。
        服务端代码路径完整实现、仅由开关短路；Batch2 实名上线时置 `on`。

        ### 状态

        已下架或已过期回 `41001`；本人查看自己的帖子不受此限。
      parameters:
        - $ref: '#/components/parameters/XInteractionId'
        - $ref: '#/components/parameters/XDeviceId'
        - $ref: '#/components/parameters/PostIdPath'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PostDetail'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '410': { $ref: '#/components/responses/Gone' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
    patch:
      tags: [post]
      summary: 编辑帖子内容
      operationId: updatePost
      x-batch: Batch2
      x-domain: post
      description: 占位。Batch2 交付，同样以 version 做乐观锁。
      parameters:
        - $ref: '#/components/parameters/PostIdPath'
      responses:
        '200': { description: 占位，Batch2 交付 }
  /posts/{post_id}/status:
    patch:
      tags: [post]
      summary: 变更帖子状态（下架 / 重新上架 / 延期）
      operationId: updatePostStatus
      x-batch: Batch1
      x-domain: post
      description: |
        ### 乐观锁强约束

        - 入参**必须携带 `version`**；缺失回 `40001`，**服务端不得兜底**
          （不允许「不传就跳过版本校验」）
        - 更新语句形如 `UPDATE post SET ... WHERE id=? AND version=?`
        - 影响行数为 0 → 回 `40903`，客户端必须强制重取详情后由用户重试

        ### 写守卫分流

        | 变更来源 | 守卫条件 |
        | --- | --- |
        | 用户意图变更（本接口） | 守 `version`（乐观锁） |
        | 系统派生变更（定时任务到期下架等） | 守 `status IN (合法前驱集)`，不守 `version` |
      parameters:
        - $ref: '#/components/parameters/XInteractionId'
        - $ref: '#/components/parameters/XDeviceId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/PostIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action, version]
              properties:
                action:
                  type: string
                  enum: [offline, republish, renew]
                  description: |
                    - `offline`：下架（active → offline）
                    - `republish`：重新上架（offline → active，重算 expire_at）
                    - `renew`：延期（active → active，顺延 expire_at）
                version:
                  type: integer
                  format: int64
                  description: 从 GET 详情或 GET /posts/mine 取得的当前版本号。**缺失回 40001**
      responses:
        '200':
          description: 变更成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          id: { type: integer, format: int64 }
                          status: { $ref: '#/components/schemas/PostStatusEnum' }
                          expire_at: { type: string, format: date-time }
                          version:
                            type: integer
                            format: int64
                            description: 变更后的新版本号
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '410': { $ref: '#/components/responses/Gone' }
        '500': { $ref: '#/components/responses/InternalError' }
  /posts/mine:
    get:
      tags: [post]
      summary: 我发布的帖子列表
      operationId: listMyPosts
      x-batch: Batch1
      x-domain: post
      description: |
        出参每项**必带 `version`**（供状态变更的乐观锁使用）。
        本人视角，`audit_status=pending/reject` 的媒体也会下发（含 `reject_reason`）。
      parameters:
        - $ref: '#/components/parameters/XInteractionId'
        - $ref: '#/components/parameters/XDeviceId'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: status
          in: query
          required: false
          description: 状态筛选，缺省返回全部
          schema: { $ref: '#/components/schemas/PostStatusEnum' }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items: { $ref: '#/components/schemas/MyPostItem' }
                          total: { type: integer, format: int32 }
                          page: { type: integer, format: int32 }
                          page_size: { type: integer, format: int32 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        单 Token 模型，有效期 30 天。请求头形如 `Authorization: Bearer <JWT>`。
        Token 失效由服务端回 `40101` 触发客户端单飞续期，
        客户端**不做本地过期判断**。

  # ===========================================================================
  # 复用参数：四个自定义请求头 + 通用分页 + 通用路径参数
  # ===========================================================================
  parameters:
    XInteractionId:
      name: X-Interaction-Id
      in: header
      required: true
      description: |
        **客户端生成**的交互标识，覆盖「缓存 → 网络 → 聚合 → 渲染」四段全链路。
        服务端**仅用于日志串联**，**严禁**用于鉴权、限流、幂等判定。
        与 `request_id` 分工不同：`request_id` 由服务端生成、是 UI 报错唯一回显值。
        `40101` 重放挂起请求时必须**沿用原值**（逐字相同）。

        ### 例外：POST /track/events

        埋点上报是本系统唯一**没有交互起点**的调用场景 —— 定时触发、进后台触发、
        登录后补报、离线回灌，四种触发方式都不属于任何进行中的用户交互；
        且一批最多 50 条事件**可能来自几十个不同交互**，让批次沿用其中任一条的
        `interaction_id` 都是错的（会把整批的排查线索指向一个无关交互）。

        因此在该接口上，本头的语义收窄为「**这一次上报传输动作**」：

        - 由上报器**为每个批次新生成**一个值，仍满足 required；
        - 事件自身所属的交互保留在**请求体** `events[].interaction_id` 字段里，
          与本头**互不覆盖**；
        - 重发（失败退回队列后重新组批）时**必须换新值**，不沿用 —— 与上面
          `40101` 重放的规则相反，因为重新组批就是一次新的传输动作。

        排查方式：查某个用户交互用事件体内的 `interaction_id`；
        查某次上报为何失败用本头的值。
      schema:
        type: string
        maxLength: 64
        example: itx_1a2b3c4d5e6f
    XDeviceId:
      name: X-Device-Id
      in: header
      required: true
      description: |
        首次启动生成的 UUID v4，落 `device.fingerprint`。
        **明确不可信**（可被清数据重置），仅作风控辅助维度之一。
        **严禁**采集 IMEI / MAC / IDFA / OAID 等设备唯一标识。
      schema:
        type: string
        format: uuid
        example: 3f2504e0-4f89-41d3-9a0c-0305e82c3301
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        **所有写接口必带**。UUID v4，36 字符，小写，带连字符。

        正则：`^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$`
        不符合回 `40001`。

        ### 四类必测输入

        | 场景 | 期望行为 |
        | --- | --- |
        | 缺失该头 | 回 `40001`，**不得放行** |
        | 大写 UUID 或 v1 UUID | 判非法，回 `40001` |
        | 重复 Key 但请求体不同 | **返回首次结果**（不按新请求体执行） |
        | 首次执行失败后用同 Key 重放 | **必须重新执行**（只有成功响应才进缓存） |

        窗口 24 小时（见 `NfrApi.idempotencyWindowHours`），
        以**服务端收到请求的时间**为准。键与首次响应体存 Redis。
        `40101` 重放挂起请求时必须**沿用原值**（逐字相同）。
        网络层重试（`50001` / `503xx` / 网络失败）同样必须**沿用原值** ——
        客户端 `HeaderInterceptor` 注入前须先判该头是否已存在，已存在则不覆盖。

        ### 例外：POST /track/events —— 每次组批必须换新 Key

        埋点批次的幂等语义与其他写接口**相反**，原因在于「重发的内容会变」：

        上报失败的批次退回队列**头部**后，会与期间新采集的事件**重新组批**，
        内容已经不是首次那一批了。而上表第三行规定「重复 Key 但请求体不同 →
        返回首次结果」，服务端会**直接命中幂等缓存返回 `code=0`，业务逻辑一行不执行**。

        于是形成这个静默丢失链条：

        ```
        第 1 次上报 50 条 → 服务端返 42906，一条未落库
        退回队列头部      → 期间新采集 20 条
        第 2 次组批       → 内容变为「原 50 条的部分 + 新 20 条」
           若沿用旧 Key   → 命中幂等缓存 → 回 code=0
                          → 客户端判成功，把 70 条全部出队
        结果：70 条事件一条未落库，且【不产生任何错误日志】
        ```

        `dropped_count` 也不会 +1（它只统计容量溢出丢弃），所有可观测性指标都是绿的
        —— 可观测性系统自己成了盲区。

        **规则：该接口每次组批生成新的 `Idempotency-Key`**（批次是「一次传输动作」，
        不是「一笔业务意图」）。

        由此带来的重复写入风险（响应丢在回程 → 同批写两次）由服务端按**事件维度**
        去重兜住，**不依赖幂等键**：埋点库唯一索引
        `uk_event_dedup(interaction_id, event_name, user_id, ts)`，
        落库用 `INSERT ... ON DUPLICATE KEY UPDATE id = id`。详见该接口的 description。
        取舍方向明确：宁可重一条（分析侧可去重），不可丢一条（永久不可恢复）。
      schema:
        type: string
        pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
        example: 9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d
    PageParam:
      name: page
      in: query
      required: false
      description: 页码，从 1 起
      schema: { type: integer, format: int32, minimum: 1, default: 1 }
    PageSizeParam:
      name: page_size
      in: query
      required: false
      description: 每页条数，默认见 NfrApi.pageSizeDefault，上限见 NfrApi.pageSizeMax
      schema: { type: integer, format: int32, minimum: 1, maximum: 50, default: 20 }
    PostIdPath:
      name: post_id
      in: path
      required: true
      description: 帖子 ID
      schema: { type: integer, format: int64, example: 1001 }

  # ===========================================================================
  # 复用响应：按 HTTP 状态码组织，每个响应列明其可能承载的全部业务 code
  # ===========================================================================
  responses:
    BadRequest:
      description: |
        `40001` 参数缺失或非法 / `40002` 未勾选用户协议与隐私政策
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          examples:
            missingParam:
              summary: 参数缺失（40001）
              value: { code: 40001, message: 缺少必填参数 grid_id, data: null, request_id: req_7f3a91 }
            notAgreed:
              summary: 未勾选协议（40002）
              value: { code: 40002, message: 请先阅读并同意用户协议与隐私政策, data: null, request_id: req_7f3a92 }
    Unauthorized:
      description: |
        `40101` 未登录或登录态失效（客户端触发单飞续期） /
        `40105` 验证码连续错误导致登录锁定 15 分钟（**携带 `Retry-After`**）
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          examples:
            expired:
              summary: 登录态失效（40101）
              value: { code: 40101, message: 登录已过期，请重新登录, data: null, request_id: req_7f3a93 }
            locked:
              summary: 登录锁定（40105）
              value: { code: 40105, message: 验证码错误次数过多，请 15 分钟后再试, data: null, request_id: req_7f3a94 }
    Forbidden:
      description: |
        `40301` 未实名今日详情额度用尽（Batch1 由开关关闭） /
        `40302` 高敏类目需先完成资质认证 /
        `40303` 该类目当前禁止发布 /
        `40304` 未实名用户发布数量已达上限 /
        `40305` 无权限访问该资源（**后台 RBAC 越权专用**，App 端接口不返回此码）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          examples:
            quotaExhausted:
              summary: 详情额度用尽（40301）
              value: { code: 40301, message: 今日可查看详情数量已用完，完成实名后不受限制, data: null, request_id: req_7f3a95 }
            needQualification:
              summary: 缺资质（40302）
              value: { code: 40302, message: 该类目需先完成资质认证, data: null, request_id: req_7f3a96 }
            categoryBanned:
              summary: 类目禁发（40303）
              value: { code: 40303, message: 该类目当前禁止发布, data: null, request_id: req_7f3a97 }
            publishLimit:
              summary: 发布上限（40304）
              value: { code: 40304, message: 未实名用户发布数量已达上限, data: null, request_id: req_7f3a98 }
    Conflict:
      description: |
        `40901` 内容命中敏感词 /
        `40902` 图片未通过审核 /
        `40903` 数据已被其他操作修改（乐观锁冲突，客户端须强制重取后由用户重试）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          examples:
            sensitiveWord:
              summary: 敏感词（40901）
              value: { code: 40901, message: 内容包含不允许发布的词汇，请修改后重试, data: null, request_id: req_7f3a99 }
            imageRejected:
              summary: 图片未过审（40902）
              value: { code: 40902, message: 图片未通过审核，请更换后重试, data: null, request_id: req_7f3a9a }
            versionConflict:
              summary: 乐观锁冲突（40903）
              value: { code: 40903, message: 数据已更新，请刷新后重试, data: null, request_id: req_7f3a9b }
    Gone:
      description: '`41001` 帖子已下架或已过期'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example: { code: 41001, message: 该信息已下架或已过期, data: null, request_id: req_7f3a9c }
    TooManyRequests:
      description: |
        `42901` AI 配额已用尽 /
        `42902` 联系次数超出每日限额 /
        `42903` 异常操作触发熔断当日冻结 /
        `42904` 推送频次受限 /
        `42905` 验证码发送过于频繁 /
        `42906` 埋点上报频次受限（**唯一不向用户呈现**，客户端静默退回队列） /
        `42907` 未登录详情浏览超限（设备 30 / IP 100 每日，任一超限即拒）

        以上 7 个 code **必须携带 `Retry-After`** 响应头，值为剩余整数秒。
        `42901`–`42905`、`42907` 客户端**不可自动重试**，须按 `Retry-After` 提示用户。
        `42907` 的提示须附登录入口，且**不告知是哪个维度超限**（风控阈值不外泄）。
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          examples:
            aiQuota:
              summary: AI 配额（42901）
              value: { code: 42901, message: 今日 AI 辅助次数已用完, data: null, request_id: req_7f3a9d }
            contactLimit:
              summary: 联系超限（42902）
              value: { code: 42902, message: 今日查看联系方式次数已达上限, data: null, request_id: req_7f3a9e }
            circuitBreak:
              summary: 熔断冻结（42903）
              value: { code: 42903, message: 操作过于频繁，该功能今日已暂停使用, data: null, request_id: req_7f3a9f }
            pushLimit:
              summary: 推送频控（42904）
              value: { code: 42904, message: 推送发送过于频繁, data: null, request_id: req_7f3aa0 }
            smsLimit:
              summary: 验证码频控（42905）
              value: { code: 42905, message: 验证码发送过于频繁，请稍后再试, data: null, request_id: req_7f3aa1 }
            trackLimit:
              summary: 埋点频控（42906，不呈现给用户）
              value: { code: 42906, message: track rate limited, data: null, request_id: req_7f3aa2 }
            guestDetailLimit:
              summary: 未登录详情浏览超限（42907）
              value: { code: 42907, message: 今日浏览次数较多，登录后可继续查看, data: null, request_id: req_7f3aa3 }
    InternalError:
      description: |
        `50001` 服务内部错误。客户端**可自动重试**，
        与 `50301`/`50302`/`50303` 及网络超时共享「总共 2 次」的重试预算，
        退避 1s / 2s 并叠加 ±20% 抖动。
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example: { code: 50001, message: 服务繁忙，请稍后重试, data: null, request_id: req_7f3aa3 }
...
  headers:
    RetryAfter:
      description: |
        剩余可重试秒数（**整数**，非 HTTP-date 形式）。
        `42901`–`42907` 与 `40105` **必须**返回本头。
      required: true
      schema: { type: integer, format: int32, minimum: 1, example: 900 }

  # ===========================================================================
  # 数据模型
  # ===========================================================================
  schemas:
...
    ApiEnvelope:
      type: object
      description: |
        统一响应包骨架。所有接口无论成功失败均返回该结构，
        各接口通过 `allOf` 收窄 `data` 的具体类型。
      required: [code, message]
      properties:
        code:
          type: integer
          format: int32
          description: 业务码。0 成功；非 0 见错误码表，前 3 位与 HTTP 状态码对齐
          example: 0
        message:
          type: string
          description: 中文文案。非 0 时可直接呈现给用户
          example: ok
        data:
          description: 业务负载。失败时恒为 null
          nullable: true
        request_id:
          type: string
          description: |
            服务端生成的请求标识，是 UI 报错时唯一回显值。
            **缓存命中时该字段不存在**，属预期行为。
          example: req_7f3a91c2
    ApiError:
      type: object
      description: 失败响应（`data` 恒为 null）
      required: [code, message, request_id]
      properties:
        code: { type: integer, format: int32, example: 40001 }
        message: { type: string, example: 缺少必填参数 grid_id }
        data: { nullable: true, example: null }
        request_id: { type: string, example: req_7f3a91c2 }

    # ---------- 通用枚举与基础类型 ----------
    PostTypeEnum:
      type: string
      description: 供需态。resource=资源方（我有），demand=需求方（我要）
      enum: [resource, demand]
    PostStatusEnum:
      type: string
      description: |
        帖子状态。`active` 在架 / `offline` 已下架 / `expired` 已过期 / `archived` 已归档。
        有效天数与自动归档天数见 `NfrPostLifecycle`。
      enum: [active, offline, expired, archived]
...
    CompletenessLevelEnum:
      type: integer
      format: int32
      description: |
        信息完整度等级，`post.completeness_level` **STORED 生成列**（不可写，
        写入报 `ERROR 3105`）。由 `required_full` / `address_precise` / `leaf_matched`
        三条件达成数映射：3→2、2→1、其他→0。
      enum: [0, 1, 2]
...
    MediaItem:
      type: object
      description: |
        帖子内的媒体项。他人视角只会出现 `audit_status=pass` 的项；
        本人视角可出现三态并携带 `reject_reason`。
      required: [media_id, url, audit_status]
      properties:
        media_id: { type: string }
        url: { type: string }
        audit_status: { $ref: '#/components/schemas/AuditStatusEnum' }
        reject_reason: { type: string, nullable: true }

    # ---------- 发布 ----------
    PostDraft:
      type: object
      description: |
        发布草稿载荷，用于 `/posts/precheck`。
        与 `PostCreateRequest` 字段一致但全部非必填（允许对半成品做预校验）。
      properties:
        type: { $ref: '#/components/schemas/PostTypeEnum' }
        leaf_category_id: { type: integer, format: int32 }
        title: { type: string, maxLength: 40 }
        price:
          type: number
          format: double
          nullable: true
          description: 价格（元）。null/缺省表示「面议」（PRD §5.8 允许价格为空）
        price_unit:
          type: string
          nullable: true
          description: 价格单位，取模板 price_units 之一；price 为 null 时无意义
        description: { type: string, maxLength: 500 }
        attributes:
          type: object
          additionalProperties: true
          description: 动态属性，键取自模板 fields[].key，落 `post.attributes` JSON 列
        lng: { type: number, format: double }
        lat: { type: number, format: double }
        address: { type: string, maxLength: 120 }
        address_precise:
          type: boolean
          description: 地址是否精确到门牌，`completeness_level` 三条件之一
        media_ids:
          type: array
          maxItems: 9
          items: { type: string }
        contact_type:
          type: string
          enum: [phone, wechat]
        contact_value: { type: string, maxLength: 64 }
    PostCreateRequest:
      allOf:
        - $ref: '#/components/schemas/PostDraft'
        - type: object
          required: [type, leaf_category_id, title, lng, lat, contact_type, contact_value]
          description: |
            正式发布载荷。派生字段
            （`l2_category_id` / `completeness_level` / `grid_id` / `expire_at` / `version`）
            **由服务端生成，客户端传了也被忽略**。
    PrecheckResult:
      type: object
      required: [passed, blocks]
      properties:
        passed:
          type: boolean
          description: true 表示无阻断项，可直接提交发布
        blocks:
          type: array
          description: 全部阻断项一次性给全（不逐条打断用户）
          items:
            type: object
            required: [code, message]
            properties:
              code:
                type: integer
                format: int32
                description: 对应的错误码（40901/40902/40303/40302/40304）
              message: { type: string, description: 可直接呈现的中文提示 }
              field:
                type: string
                nullable: true
                description: 关联字段名，供客户端定位到具体表单项
        completeness_level:
          $ref: '#/components/schemas/CompletenessLevelEnum'
        derived:
          type: object
          description: 服务端预演算的派生值，供客户端提前展示完整度提示
          properties:
            required_full: { type: boolean }
            address_precise: { type: boolean }
            leaf_matched: { type: boolean }
    PostDetail:
      type: object
      description: |
        帖子详情。**出参必带 `version`**（供 `PATCH /posts/{id}/status` 乐观锁入参）。
        **不含完整联系方式**，仅 `contact_mask`；完整值只由
        `GET /posts/{id}/contact` 返回。
      required: [id, type, leaf_category_id, l2_category_id, title, status, version, completeness_level]
      properties:
        id: { type: integer, format: int64, example: 1001 }
        type: { $ref: '#/components/schemas/PostTypeEnum' }
        leaf_category_id: { type: integer, format: int32, example: 10101 }
        l2_category_id:
          type: integer
          format: int32
          description: STORED 生成列 `leaf_category_id DIV 100`，不可写
          example: 101
        category_path:
          type: array
          description: 类目面包屑名称，L1 → L2 → L3
          items: { type: string }
        title: { type: string }
        price: { type: number, format: double, nullable: true, description: 价格（元）；null=面议 }
        price_unit: { type: string, nullable: true, description: 价格单位；price 为 null 时无意义 }
        description: { type: string, nullable: true }
        attributes:
          type: object
          additionalProperties: true
          description: 动态属性，来自 `post.attributes` JSON 列
        lng: { type: number, format: double, description: GCJ-02，5 位小数 }
        lat: { type: number, format: double, description: GCJ-02，5 位小数 }
        address: { type: string, nullable: true }
        distance_m:
          type: integer
          format: int32
          nullable: true
          description: 与请求方视野中心的距离（米）
        media:
          type: array
          description: 他人视角仅含 audit_status=pass 的项
          items: { $ref: '#/components/schemas/MediaItem' }
        contact_mask:
          type: string
          description: 脱敏联系方式。完整值只由 /posts/{id}/contact 返回
          example: 138****8000
        completeness_level: { $ref: '#/components/schemas/CompletenessLevelEnum' }
        status: { $ref: '#/components/schemas/PostStatusEnum' }
        publish_at: { type: string, format: date-time }
        expire_at:
          type: string
          format: date-time
          description: 到期时间。倒计时文案允许用本地时间计算
        version:
          type: integer
          format: int64
          description: 乐观锁版本号。**PATCH status 时必须原样回传**
        author:
          $ref: '#/components/schemas/AuthorBrief'
    AuthorBrief:
      type: object
      description: |
        作者摘要，严格遵循 `user` 对外视图白名单。
        `phone_mask` / `real_name_enc` / `id_card_hash` **不出现**。
      required: [id, nickname, realname_status]
      properties:
        id: { type: integer, format: int64 }
        nickname: { type: string }
        avatar_url: { type: string, nullable: true }
        realname_status: { $ref: '#/components/schemas/RealnameStatusEnum' }
        qualification_badges:
          type: array
          items: { type: string }
    MyPostItem:
      type: object
      description: 我的帖子列表项。**必带 `version`**；本人视角下发媒体三态
      required: [id, type, title, status, version, completeness_level]
      properties:
        id: { type: integer, format: int64 }
        type: { $ref: '#/components/schemas/PostTypeEnum' }
        leaf_category_id: { type: integer, format: int32 }
        title: { type: string }
        cover_media:
          $ref: '#/components/schemas/MediaItem'
        completeness_level: { $ref: '#/components/schemas/CompletenessLevelEnum' }
        status: { $ref: '#/components/schemas/PostStatusEnum' }
        publish_at: { type: string, format: date-time }
        expire_at: { type: string, format: date-time }
        view_count: { type: integer, format: int32 }
        contact_count:
          type: integer
          format: int32
          description: 被查看联系方式的次数，来自 contact_event 聚合
        version: { type: integer, format: int64 }
...